docs: Add files used and credited link
Updated the list of the four files used in running the project
in the repo.

The main python bikeshare.py file pulls data from one of the .csv
files based on the user's choice of data request:

- Chicago

- Washington

- New York City

The data is then pulled into the functions in the python code
to run statististical usage of bike rentals.

Finally, these results are given to the user and the user is
asked if they would like to run another query.