style: Add text to the resources used in the project
There is a list of resources used in the development of the
bikeshare project at the end of the bikeshare.py file.

A note was added at the beginning of the section to let the
reader know that there is another file, a .txt file, that also
has a list of the used resources.