refactor: Remove canned test code for user input
In the bikeshare.py file, the code/project gets input from the user
to use to filter the bikeshare data.

During development of the project, after the code was working,
a canned user input was used to help speed up testing of new
code/code functions rather than having to manually input those
inputs for each test run of the code.

The code was commented out but should have been removed.

This commit removes that canned code.